\name{latticles2}
\alias{latticles2}
\title{Launch latticles2 GUI}

\usage{
latticles2()
}

\description{
Importe artigos de um grupo de pesquisadores, classifique pelo Qualis-CAPES (novo), 
contabilize producoes tecnicas e mais, tudo de forma automatizada a partir de arquivos XML do curriculo Lattes.
}

\details{
\tabular{ll}{
Package: \tab latticles2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-02-15\cr
License: \tab GPL (>= 2)\cr
}
}

\note{
\emph{latticles2}, um projeto em continuo desenvolvimento para avaliacao eficiente de producoes cientificas.
Criticas, comentarios e sugestoes sao sempre bem-vindas.
}

\author{
Anderson Rodrigo da Silva (Instituto Federal Goiano - Campus Urutai)

Maintainer: Anderson Rodrigo da Silva <anderson.silva@ifgoiano.edu.br>
}

\examples{
\dontrun{
latticles2()
}
}
\keyword{latticles2}
